% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawDateQuizData.R
\name{addLabels}
\alias{addLabels}
\title{Add the labels to files' data fields from the associated dictionary files}
\usage{
addLabels(files, dictionaryPrefix = "")
}
\arguments{
\item{files}{list of files to add labels to}

\item{dictionaryPrefix}{prefix to ensure the correct dictionary is found}
}
\value{
files with labels added
}
\description{
Add the labels to files' data fields from the associated dictionary files
}
