% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_responses.R
\name{annotate_responses.AdvisedTrial.binary}
\alias{annotate_responses.AdvisedTrial.binary}
\title{Add variables assessing the correctness, influence etc. of estimates}
\usage{
annotate_responses.AdvisedTrial.binary(AdvisedTrial)
}
\arguments{
\item{AdvisedTrial}{tbl of initial estimates and final decisions made with
advice in the dates task}
}
\value{
\code{AdvisedTrial} with variables:

\itemize{
\item{[reponse/advisor#]Correct[Final]}{whether the chosen side was correct}
\item{EstimateIncrease}{Difference between final and initial estimate}
\item{Influence}{Difference between final and initial estimate signed in the direction of the advice}
\item{InfluenceCapped}{Influence capped by the maximum potential increase}
}
}
\description{
Add variables assessing the correctness, influence etc. of estimates
}
