% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_responses.R
\name{annotate_responses.AdvisedTrial}
\alias{annotate_responses.AdvisedTrial}
\title{Add variables assessing the correctness, influence etc. of estimates}
\usage{
annotate_responses.AdvisedTrial(AdvisedTrial)
}
\arguments{
\item{AdvisedTrial}{tbl of initial estimates and final decisions made with
advice in the dates task}
}
\value{
\code{AdvisedTrial} with variables:

\itemize{
\item{[reponse/advisor#]Correct[Final]}{whether the chosen side was correct}
\item{[reponse/advisor#]Error[Final]}{absolute difference between the correct answer and the centre of the chosen side}
\item{EstimateIncrease}{Difference between final and initial estimate}
\item{Influence}{Difference between final and initial estimate signed in the direction of the advice}
\item{WOAraw}{Weight on advice = (final - initial) / (advice - initial)}
\item{WOA}{Winzorized version of WoA Raw to range [0-1]}
}
}
\description{
Add variables assessing the correctness, influence etc. of estimates
}
