% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_responses.R
\name{annotate_responses}
\alias{annotate_responses}
\title{Add variables assessing the correctness, influence etc. of estimates}
\usage{
annotate_responses(AdvisedTrial)
}
\arguments{
\item{AdvisedTrial}{tbl of initial estimates and final decisions made with
advice}
}
\value{
\code{AdvisedTrial} with variables:

\itemize{
\item{[reponse/advisor#]Correct[Final]}{whether the estimate was correct}
\item{[reponse/advisor#]Error[Final]}{absolute difference between the correct answer and the centre of the estimate}
}
}
\description{
Add variables assessing the correctness, influence etc. of estimates
}
