% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_responses.R
\name{annotate_responses.trials}
\alias{annotate_responses.trials}
\title{Add variables assessing the correctness, influence etc. of estimates}
\usage{
annotate_responses.trials(trials, max_conf = 50)
}
\arguments{
\item{trials}{tbl of initial estimates and final decisions made with
advice in the dots task}

\item{max_conf}{maximum confidence rating on the scale}
}
\value{
\code{AdvisedTrial} with variables:

\itemize{
\item{[reponse/advisor#]Correct[Final]}{whether the estimate was correct}
\item{estimateIncrease}{Difference between final and initial estimate}
\item{adviceInfluence}{Difference between final and initial estimate signed in the direction of the advice}
\item{adviceInfluenceCapped}{Difference between final and initial estimate signed in the direction of the advice capped to the maximum value which could have been given if confidence had increased and the answer side remained the same.}
}
}
\description{
Add variables assessing the correctness, influence etc. of estimates
}
