% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawDateQuizData.R
\name{dictToJSON}
\alias{dictToJSON}
\title{Convert a dictionary CSV file into a metadata JSON file}
\usage{
dictToJSON(dict)
}
\arguments{
\item{dict}{dictionary.csv to convert}
}
\description{
Convert a dictionary CSV file into a metadata JSON file
}
