% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_core.R
\name{.lookupAdvisorProperty}
\alias{.lookupAdvisorProperty}
\title{Return a property of an Advisor from their ID for a given participant.}
\usage{
.lookupAdvisorProperty(property, advisorId, participantId, advisors)
}
\arguments{
\item{property}{to return. Should be a column in advisors}

\item{advisorId}{id of the advisor}

\item{participantId}{will be joined to advisorId as a data.frame, so must be of the same length}

\item{advisors}{data frame of advisors}
}
\value{
property of specified advisor
}
\description{
Return a property of an Advisor from their ID for a given participant.
}
