% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawDateQuizData.R
\name{exportFullData}
\alias{exportFullData}
\title{Export data complete with tags and labels}
\usage{
exportFullData(files, name = "datequiz.full")
}
\description{
Export data complete with tags and labels
}
