% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawDateQuizData.R
\name{fetchRawData}
\alias{fetchRawData}
\title{Fetch raw data from the project server}
\usage{
fetchRawData()
}
\arguments{
\item{fileDetails}{tbl row describing the file with $url, $study, $version, $table}
}
\value{
tibble of parsed data
}
\description{
Fetch raw data from the project server
}
