% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_core.R
\name{findAdvisorGroup}
\alias{findAdvisorGroup}
\title{Return an Advisor's groupId from their ID for a given participant. Updated version of getAdviceType}
\usage{
findAdvisorGroup(advisorId, participantId, advisors)
}
\arguments{
\item{advisorId}{id of the advisor}

\item{participantId}{will be joined to advisorId as a data.frame, so must be of the same length}

\item{advisors}{data frame of advisors}
}
\value{
adviceType of specified advisor
}
\description{
Return an Advisor's groupId from their ID for a given participant. Updated version of getAdviceType
}
