% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_core.R
\name{findInfluence}
\alias{findInfluence}
\title{Return a vector of influences of advisors. Influence is +confidenceShift
where the advisor agrees, and -confidenceShift where the advisor disagrees.}
\usage{
findInfluence(advisorAgreements, confidenceShift)
}
\arguments{
\item{advisorAgreements}{logical vector of whether the advisor agreed with the participant}

\item{confidenceShifts}{the parameters are all bound together into a
dataframe so must all be the same length}
}
\value{
vector of influences of the advisors. NA where advisorId is NA
}
\description{
Return a vector of influences of advisors. Influence is +confidenceShift
where the advisor agrees, and -confidenceShift where the advisor disagrees.
}
