% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_responses.R
\name{get_influence_binary}
\alias{get_influence_binary}
\title{Calculate influence for binary advised trials}
\usage{
get_influence_binary(AdvisedTrial)
}
\arguments{
\item{AdvisedTrial}{tbl of initial estimates and final decisions made with
advice}
}
\value{
\code{AdvisedTrial} with columns appended containing influence
  calculations
}
\description{
Calculate influence for binary advised trials
}
\details{
Adjusted weight on advice (woa) is capped by limiting the maximum
  change where answers are reduced to be equivalent to the maximum possible
  change if answers had been increased. This compensates for the fact that,
  once the initial estimate marks a point to one side of a scale, there is by
  definition a greater room to move towards the other side of the scale than
  there is to increase the rating on the chosen side.

  The returned tbl has the following additional columns for each advisor
  (\code{AdvisedTrial$advisor#...}):

  \itemize{
  \item{changedOfMind}{whether the chosen side is the same for the initial estimate and final decision}
  \item{increase}{how much confidence has increased from initial estimate to final decision}
  \item{agrees}{whether the advisor agrees with the initial decision}
  \item{woaRaw}{how much the}
  }
}
