% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_core.R
\name{getConfidenceShift}
\alias{getConfidenceShift}
\title{Find the confidence shift in a given trial}
\usage{
getConfidenceShift(t, rawShift = FALSE, forceRecalculate = FALSE)
}
\arguments{
\item{t}{trial list}

\item{rawShift}{whether to report the confidence shift without adjusting for the assymetric scale}

\item{forceRecalulate}{if true, simply return the appropriate column from t if it exists already}
}
\value{
a vector of confidence shifts for trial list t
}
\description{
Find the confidence shift in a given trial
}
