% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawDateQuizData.R
\name{getDictionaries.dates}
\alias{getDictionaries.dates}
\title{Fetch dictionary files from the local webserver}
\usage{
getDictionaries.dates()
}
\value{
list of dictionaries in JSON format
}
\description{
Fetch dictionary files from the local webserver
}
