% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM_core.R
\name{loadFilesFromFolder}
\alias{loadFilesFromFolder}
\title{Fetch experiment JSON files from a folder}
\usage{
loadFilesFromFolder(folderName)
}
\arguments{
\item{folderName}{path to folder to process}
}
\value{
list of tables of data extracted from JSONs
}
\description{
Fetch experiment JSON files from a folder
}
