% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_responses.R
\name{mark_responses.binary}
\alias{mark_responses.binary}
\title{Add variables indicating error, correctness, etc. for estimates.}
\usage{
mark_responses.binary(AdvisedTrial)
}
\arguments{
\item{AdvisedTrial}{tbl of initial estimates and final decisions made with
advice}
}
\value{
\code{AdvisedTrial} with confidenceScore and correct columns for initial estimate, final decision, and any advisory estimates
}
\description{
Add variables indicating error, correctness, etc. for estimates.
}
\details{
confidenceScore is a confidence expressed as a proportion of maximum confidence
}
