% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_responses.R
\name{mark_responses}
\alias{mark_responses}
\title{Add variables indicating error, correctness, etc. for estimates}
\usage{
mark_responses(AdvisedTrial)
}
\arguments{
\item{AdvisedTrial}{tbl of initial estimates and final decisions made with
advice}
}
\value{
\code{AdvisedTrial} with correct and error columns for initial estimate, final decision, and any advisory estimates
}
\description{
Add variables indicating error, correctness, etc. for estimates
}
