% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_responses.R
\name{rate_influence.binary}
\alias{rate_influence.binary}
\title{Add variables indicating error, correctness, etc. for estimates}
\usage{
rate_influence.binary(AdvisedTrial)
}
\arguments{
\item{AdvisedTrial}{tbl of initial estimates and final decisions made with
advice}
}
\value{
\code{AdvisedTrial} with additional columns \code{advisor#...} where ... is:
\itemize{
\item{EstimateIncrease}{Difference between final and initial estimate}
\item{Influence}{Difference between final and initial estimate signed in the direction of the advice}
\item{InfluenceCapped}{Influence capped by the maximum potential increase}
}
}
\description{
Add variables indicating error, correctness, etc. for estimates
}
