% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_responses.R
\name{rate_influence}
\alias{rate_influence}
\title{Add variables indicating error, correctness, etc. for estimates}
\usage{
rate_influence(AdvisedTrial)
}
\arguments{
\item{AdvisedTrial}{tbl of initial estimates and final decisions made with
advice}
}
\value{
\code{AdvisedTrial} with additional columns \code{advisor#...} where ... is:
\itemize{
\item{estimateIncrease}{Difference between final and initial estimate}
\item{influence}{Difference between final and initial estimate signed in the direction of the advice}
\item{woaRaw}{Weight on advice = (final - initial) / (advice - initial)}
\item{woa}{Winzorized version of WoA Raw to range [0-1]}
}
}
\description{
Add variables indicating error, correctness, etc. for estimates
}
