% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawDateQuizData.R
\name{saveLocalDictionaries}
\alias{saveLocalDictionaries}
\title{Save local copy of dictionaries}
\usage{
saveLocalDictionaries(dicts, prefix = "")
}
\arguments{
\item{dicts}{tbl of dictionaries with columns name and json}

\item{prefix}{string to include before the dictionary name}
}
\value{
NULL (invisible)
}
\description{
Save local copy of dictionaries
}
