% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_experiment.R
\name{select_experiment}
\alias{select_experiment}
\title{Expand nested tables into their own tibbles in the environment}
\usage{
select_experiment(
  project,
  f = function(x) dlpyr::filter(x),
  envir = .GlobalEnv
)
}
\arguments{
\item{project}{'dotstask' or 'datequiz'}

\item{f}{filter function like \code{function(x) dplyr::filter(x, ...)}
specifying filter conditions}

\item{envir}{environment to write to}
}
\description{
Expand nested tables into their own tibbles in the environment
}
