% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tada.R
\name{tada}
\alias{tada}
\title{Wrapper around utils::data() which tibblifies the resulting list objects}
\usage{
tada(list, envir = .GlobalEnv, overwrite = T, ...)
}
\arguments{
\item{list}{character vector of sources to load}

\item{envir}{environment to write to}

\item{overwrite}{whether to overwrite existing objects in \code{envir}}

\item{...}{
  Arguments passed on to \code{\link[utils:data]{utils::data}}
  \describe{
    \item{\code{package}}{
    a character vector giving the package(s) to look
    in for data sets, or \code{NULL}.

    By default, all packages in the search path are used, then
    the \file{data} subdirectory (if present) of the current working
    directory.
  }
    \item{\code{lib.loc}}{a character vector of directory names of \R libraries,
    or \code{NULL}.  The default value of \code{NULL} corresponds to all
    libraries currently known.}
    \item{\code{verbose}}{a logical.  If \code{TRUE}, additional diagnostics are
    printed.}
  }}
}
\value{
tibble report of variables loaded and their statuses (skipped,
  tibblified, or imported)
}
\description{
Wrapper around utils::data() which tibblifies the resulting list objects
}
