% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawDotsTaskData.R
\name{tagData.dots}
\alias{tagData.dots}
\title{Add information about studies to the raw data table}
\usage{
tagData.dots(files)
}
\arguments{
\item{files}{raw data table}
}
\value{
files with metadata added
}
\description{
Add information about studies to the raw data table
}
