% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawDateQuizData.R
\name{tagData}
\alias{tagData}
\title{Add information about studies to the raw data table}
\usage{
tagData(files)
}
\arguments{
\item{files}{raw data table}
}
\value{
files with metadata added
}
\description{
Add information about studies to the raw data table
}
